AI Detection Code
Script class AIDetection takes Range, FOV and a Target as a parameters,
if Target is within range and the FOV cone, a raycast then ensures that
no colliders are in the way.  The usual quick knock-up caveats apply,
like the range check only considers the middle of the game object, and
the fact that the hit detection part does nothing.

I've added three debug lines, one is the 'forward' direction, one for
the Target check and the line then turns blue if a detection has
occurred.