Playing with creating objects
The problem of creating distinct data for two view in the same scene is
lingering.
The current approach I'm using is to create an object that can
instantiate other objects depending on view. This controlling object
should handle matching up network IDs and such.
I'm not sure this works, but I'll continue titting about.