Spawn point added
This should be a separate component on the level, but it'll do for now.

I've also taken the RigidBody component off the walls as they're static
colliders and have no need for such tomfoolery. I found this out as
Unity shit itself in a bloody typhoon of PhysX-related warnings.

According to the commit-log, I have done other things as well. Mostly
just twiddling, I'm sure.