Physics twiddling
I've assigned geometry collider a separate layer. This layer is set not
to collide with itself, so the pairs problem should be non-existent now.

However, in the process of testing this it seems the collider-creation
is incredibly slow due to the use of wall edges as colliders. Given that
I have no desire for the walls to be hollow, I might as well create box
colliders per-wall rather than per-edge, but my brain isn't up to
thinking about how to do that just now.