Visibility region added
I've made a flaky script for rendering the area visible to the player.
It's slow and ugly, as usual.
Having made this, I have become more aware that I quite dislike such
visibility schemes. The effect is clumsy, blinkered, and makes
forward-planning horrible.
Anyway, the code is there, even if it does nothing for gameplay.