Horrible synchronisation things.
Right. All sorts of toss here, though hardly any of value.
A clunky first pass at object synchronisation. Scripts need to be added
that respond to serialised data, but the crux of this is that each
spawned object has a view-agnostic controller script that knows how to
poke each view when shared data is changed.
I got bored on Saturday and made the Admin-view hideous for no
particular reason. Sorry.
There's a seriously ghetto adjustment to the AIWander script that just
fluffs the AI towards a player if spotted.
The player has a health value and the game will end if it hits zero.
Overlapping the AI lowers it.
NPCs are very temporarily turned off, as they kept murdering me while I
was debugging things and they need converting to be level-objects
anyway.