Initial camera placement level generator stage plus all the supporting cast needed to make this work.
At the moment it's a bit noddy, but I thought I'll check it in now to avoid merge fails down the road.  All it does is place the cameras in one of the randomly chosen corners of the room.
Each room has a random chance of having a camera placed in it at all.  After this commit it'll be easier to them hook up the security system.